Add diagonal dataset
The diagonal dataset contains positive data points below the main diagonal and negative data points above. The positive training data lies in a subsection of the lower triangle, while the postiive test data is distributed throughout the entire lower triangle. Unlike previous datasets it provides an example of differing training and testing data and can demonstrate the concept of overfitting.